Fix warnings and build static link libraries

+++ Remc/src/Remc/ImGui/ImGuiLayer.cpp
@@ -5,23 +5,21 @@
 #include "examples/imgui_impl_glfw.h"
 #include "examples/imgui_impl_opengl3.h"
 
-#include "GLFW/glfw3.h"
-#include "glad/glad.h"
-
 #include "Remc/Application.h"
 
-namespace Remc
-{
+// TEMPORARY
+#include <GLFW/glfw3.h>
+#include <glad/glad.h>
+
+namespace Remc {
 
 	ImGuiLayer::ImGuiLayer()
 		: Layer("ImGuiLayer")
 	{
-
 	}
 
 	ImGuiLayer::~ImGuiLayer()
 	{
-
 	}
 
 	void ImGuiLayer::OnAttach()
@@ -54,7 +52,6 @@ namespace Remc
 
 		// Setup Platform/Renderer bindings
 		ImGui_ImplGlfw_InitForOpenGL(window, true);
-
 		ImGui_ImplOpenGL3_Init("#version 410");
 	}
 
@@ -64,7 +61,7 @@ namespace Remc
 		ImGui_ImplGlfw_Shutdown();
 		ImGui::DestroyContext();
 	}
-
+	
 	void ImGuiLayer::Begin()
 	{
 		ImGui_ImplOpenGL3_NewFrame();
@@ -76,10 +73,12 @@ namespace Remc
 	{
 		ImGuiIO& io = ImGui::GetIO();
 		Application& app = Application::Get();
-		io.DisplaySize = ImVec2(app.GetWindow().GetWidth(), app.GetWindow().GetHeight());
+		io.DisplaySize = ImVec2((float)app.GetWindow().GetWidth(), (float)app.GetWindow().GetHeight());
+
 		// Rendering
 		ImGui::Render();
 		ImGui_ImplOpenGL3_RenderDrawData(ImGui::GetDrawData());
+
 		if (io.ConfigFlags & ImGuiConfigFlags_ViewportsEnable)
 		{
 			GLFWwindow* backup_current_context = glfwGetCurrentContext();
